mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
package nextstep.mvc.controller.tobe;

import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.lang.reflect.InvocationTargetException;
import java.lang.reflect.Method;
import nextstep.mvc.view.ModelAndView;

public class HandlerExecution {

    private final Method method;

    public HandlerExecution(final Method method) {
        this.method = method;
    }

    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) {
        final Class<?> declaringClass = method.getDeclaringClass();
        final Object clazz;
        try {
            clazz = declaringClass.getDeclaredConstructor()
                    .newInstance();
            return (ModelAndView) method.invoke(clazz, new Object[]{request, response});
        } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
            throw new RuntimeException(e);
        }
    }
}
